added xcase tests
