+++ lang/core/src/syntax/consumer/mod.rs
@@ -65,186 +65,3 @@ impl Subst for Consumer {
         }
     }
 }
-
-#[cfg(test)]
-mod consumer_tests {
-    use crate::{
-        syntax::{
-            context::ContextBinding, statement::Cut, types::Ty, Case, Clause, Consumer, Covar,
-            Covariable, Producer, Var, Variable,
-        },
-        traits::{free_vars::FreeV, substitution::Subst},
-    };
-    use std::{collections::HashSet, rc::Rc};
-
-    fn example_case() -> Consumer {
-        Case {
-            cases: vec![
-                Clause {
-                    xtor: "Nil".to_owned(),
-                    context: vec![],
-                    rhs: Rc::new(
-                        Cut {
-                            producer: Rc::new(
-                                Variable {
-                                    var: "x".to_owned(),
-                                }
-                                .into(),
-                            ),
-                            consumer: Rc::new(
-                                Covariable {
-                                    covar: "a".to_owned(),
-                                }
-                                .into(),
-                            ),
-                        }
-                        .into(),
-                    ),
-                },
-                Clause {
-                    xtor: "Cons".to_owned(),
-                    context: vec![
-                        ContextBinding::VarBinding {
-                            var: "x".to_owned(),
-                            ty: Ty::Int(),
-                        },
-                        ContextBinding::VarBinding {
-                            var: "xs".to_owned(),
-                            ty: Ty::Decl("ListInt".to_owned()),
-                        },
-                        ContextBinding::CovarBinding {
-                            covar: "a".to_owned(),
-                            ty: Ty::Int(),
-                        },
-                    ],
-                    rhs: Rc::new(
-                        Cut {
-                            producer: Rc::new(
-                                Variable {
-                                    var: "x".to_owned(),
-                                }
-                                .into(),
-                            ),
-                            consumer: Rc::new(
-                                Covariable {
-                                    covar: "a".to_owned(),
-                                }
-                                .into(),
-                            ),
-                        }
-                        .into(),
-                    ),
-                },
-            ],
-        }
-        .into()
-    }
-
-    fn example_prodsubst() -> Vec<(Producer, Var)> {
-        vec![(
-            Variable {
-                var: "y".to_owned(),
-            }
-            .into(),
-            "x".to_owned(),
-        )]
-    }
-    fn example_conssubst() -> Vec<(Consumer, Covar)> {
-        vec![(
-            Covariable {
-                covar: "b".to_owned(),
-            }
-            .into(),
-            "a".to_owned(),
-        )]
-    }
-
-    #[test]
-    fn display_case() {
-        let result = format!("{}", example_case());
-        let expected =
-            "case { Nil() => <x | 'a>, Cons(x : Int, xs : ListInt, 'a :cnt Int) => <x | 'a> }"
-                .to_owned();
-        assert_eq!(result, expected)
-    }
-
-    #[test]
-    fn free_vars_case() {
-        let result = example_case().free_vars();
-        let expected = HashSet::from(["x".to_owned()]);
-        assert_eq!(result, expected)
-    }
-
-    #[test]
-    fn free_covars_case() {
-        let result = example_case().free_covars();
-        let expected = HashSet::from(["a".to_owned()]);
-        assert_eq!(result, expected)
-    }
-
-    #[test]
-    fn subst_case() {
-        let result = example_case().subst_sim(&example_prodsubst(), &example_conssubst());
-        let expected = Case {
-            cases: vec![
-                Clause {
-                    xtor: "Nil".to_owned(),
-                    context: vec![],
-                    rhs: Rc::new(
-                        Cut {
-                            producer: Rc::new(
-                                Variable {
-                                    var: "y".to_owned(),
-                                }
-                                .into(),
-                            ),
-                            consumer: Rc::new(
-                                Covariable {
-                                    covar: "b".to_owned(),
-                                }
-                                .into(),
-                            ),
-                        }
-                        .into(),
-                    ),
-                },
-                Clause {
-                    xtor: "Cons".to_owned(),
-                    context: vec![
-                        ContextBinding::VarBinding {
-                            var: "x0".to_owned(),
-                            ty: Ty::Int(),
-                        },
-                        ContextBinding::VarBinding {
-                            var: "x1".to_owned(),
-                            ty: Ty::Decl("ListInt".to_owned()),
-                        },
-                        ContextBinding::CovarBinding {
-                            covar: "a0".to_owned(),
-                            ty: Ty::Int(),
-                        },
-                    ],
-                    rhs: Rc::new(
-                        Cut {
-                            producer: Rc::new(
-                                Variable {
-                                    var: "x0".to_owned(),
-                                }
-                                .into(),
-                            ),
-                            consumer: Rc::new(
-                                Covariable {
-                                    covar: "a0".to_owned(),
-                                }
-                                .into(),
-                            ),
-                        }
-                        .into(),
-                    ),
-                },
-            ],
-        }
-        .into();
-        assert_eq!(result, expected)
-    }
-}
