+++ lang/core/src/syntax/producer/mod.rs
@@ -71,177 +71,3 @@ impl Subst for Producer {
         }
     }
 }
-
-#[cfg(test)]
-mod producer_tests {
-    use crate::{
-        syntax::{
-            context::ContextBinding, statement::Cut, types::Ty, Clause, Cocase, Consumer, Covar,
-            Covariable, Producer, Var, Variable,
-        },
-        traits::{free_vars::FreeV, substitution::Subst},
-    };
-    use std::{collections::HashSet, rc::Rc};
-
-    fn example_cocase() -> Producer {
-        Cocase {
-            cocases: vec![
-                Clause {
-                    xtor: "Fst".to_owned(),
-                    context: vec![
-                        ContextBinding::VarBinding {
-                            var: "x".to_owned(),
-                            ty: Ty::Int(),
-                        },
-                        ContextBinding::CovarBinding {
-                            covar: "a".to_owned(),
-                            ty: Ty::Int(),
-                        },
-                    ],
-                    rhs: Rc::new(
-                        Cut {
-                            producer: Rc::new(
-                                Variable {
-                                    var: "x".to_owned(),
-                                }
-                                .into(),
-                            ),
-                            consumer: Rc::new(
-                                Covariable {
-                                    covar: "a".to_owned(),
-                                }
-                                .into(),
-                            ),
-                        }
-                        .into(),
-                    ),
-                },
-                Clause {
-                    xtor: "Snd".to_owned(),
-                    context: vec![],
-                    rhs: Rc::new(
-                        Cut {
-                            producer: Rc::new(
-                                Variable {
-                                    var: "x".to_owned(),
-                                }
-                                .into(),
-                            ),
-                            consumer: Rc::new(
-                                Covariable {
-                                    covar: "a".to_owned(),
-                                }
-                                .into(),
-                            ),
-                        }
-                        .into(),
-                    ),
-                },
-            ],
-        }
-        .into()
-    }
-
-    fn example_prodsubst() -> Vec<(Producer, Var)> {
-        vec![(
-            Variable {
-                var: "y".to_owned(),
-            }
-            .into(),
-            "x".to_owned(),
-        )]
-    }
-    fn example_conssubst() -> Vec<(Consumer, Covar)> {
-        vec![(
-            Covariable {
-                covar: "b".to_owned(),
-            }
-            .into(),
-            "a".to_owned(),
-        )]
-    }
-
-    #[test]
-    fn display_cocase() {
-        let result = format!("{}", example_cocase());
-        let expected =
-            "cocase { Fst(x : Int, 'a :cnt Int) => <x | 'a>, Snd() => <x | 'a> }".to_owned();
-        assert_eq!(result, expected)
-    }
-
-    #[test]
-    fn free_vars_cocase() {
-        let result = example_cocase().free_vars();
-        let expected = HashSet::from(["x".to_owned()]);
-        assert_eq!(result, expected)
-    }
-
-    #[test]
-    fn free_covars_cocase() {
-        let result = example_cocase().free_covars();
-        let expected = HashSet::from(["a".to_owned()]);
-        assert_eq!(result, expected)
-    }
-
-    #[test]
-    fn subst_cocase() {
-        let result = example_cocase().subst_sim(&example_prodsubst(), &example_conssubst());
-        let expected = Cocase {
-            cocases: vec![
-                Clause {
-                    xtor: "Fst".to_owned(),
-                    context: vec![
-                        ContextBinding::VarBinding {
-                            var: "x0".to_owned(),
-                            ty: Ty::Int(),
-                        },
-                        ContextBinding::CovarBinding {
-                            covar: "a0".to_owned(),
-                            ty: Ty::Int(),
-                        },
-                    ],
-                    rhs: Rc::new(
-                        Cut {
-                            producer: Rc::new(
-                                Variable {
-                                    var: "x0".to_owned(),
-                                }
-                                .into(),
-                            ),
-                            consumer: Rc::new(
-                                Covariable {
-                                    covar: "a0".to_owned(),
-                                }
-                                .into(),
-                            ),
-                        }
-                        .into(),
-                    ),
-                },
-                Clause {
-                    xtor: "Snd".to_owned(),
-                    context: vec![],
-                    rhs: Rc::new(
-                        Cut {
-                            producer: Rc::new(
-                                Variable {
-                                    var: "y".to_owned(),
-                                }
-                                .into(),
-                            ),
-                            consumer: Rc::new(
-                                Covariable {
-                                    covar: "b".to_owned(),
-                                }
-                                .into(),
-                            ),
-                        }
-                        .into(),
-                    ),
-                },
-            ],
-        }
-        .into();
-        assert_eq!(result, expected)
-    }
-}
