+++ lang/core/src/syntax/term/xcase.rs
@@ -41,3 +41,185 @@ impl<T: PrdCns> From<XCase<T>> for Term<T> {
         Term::XCase(value)
     }
 }
+
+#[cfg(test)]
+mod xcase_tests {
+    use super::{FreeV, XCase};
+    use crate::syntax::{
+        context::ContextBinding,
+        statement::Cut,
+        term::{Cns, Prd},
+        types::Ty,
+        Clause, Covariable, Variable,
+    };
+    use std::{collections::HashSet, rc::Rc};
+
+    fn example_cocase() -> XCase<Prd> {
+        XCase {
+            prdcns: Prd,
+            clauses: vec![
+                Clause {
+                    xtor: "Fst".to_owned(),
+                    context: vec![
+                        ContextBinding::VarBinding {
+                            var: "x".to_owned(),
+                            ty: Ty::Int(),
+                        },
+                        ContextBinding::CovarBinding {
+                            covar: "a".to_owned(),
+                            ty: Ty::Int(),
+                        },
+                    ],
+                    rhs: Rc::new(
+                        Cut {
+                            producer: Rc::new(
+                                Variable {
+                                    var: "x".to_owned(),
+                                }
+                                .into(),
+                            ),
+                            consumer: Rc::new(
+                                Covariable {
+                                    covar: "a".to_owned(),
+                                }
+                                .into(),
+                            ),
+                        }
+                        .into(),
+                    ),
+                },
+                Clause {
+                    xtor: "Snd".to_owned(),
+                    context: vec![],
+                    rhs: Rc::new(
+                        Cut {
+                            producer: Rc::new(
+                                Variable {
+                                    var: "x".to_owned(),
+                                }
+                                .into(),
+                            ),
+                            consumer: Rc::new(
+                                Covariable {
+                                    covar: "a".to_owned(),
+                                }
+                                .into(),
+                            ),
+                        }
+                        .into(),
+                    ),
+                },
+            ],
+        }
+        .into()
+    }
+
+    fn example_case() -> XCase<Cns> {
+        XCase {
+            prdcns: Cns,
+            clauses: vec![
+                Clause {
+                    xtor: "Nil".to_owned(),
+                    context: vec![],
+                    rhs: Rc::new(
+                        Cut {
+                            producer: Rc::new(
+                                Variable {
+                                    var: "x".to_owned(),
+                                }
+                                .into(),
+                            ),
+                            consumer: Rc::new(
+                                Covariable {
+                                    covar: "a".to_owned(),
+                                }
+                                .into(),
+                            ),
+                        }
+                        .into(),
+                    ),
+                },
+                Clause {
+                    xtor: "Cons".to_owned(),
+                    context: vec![
+                        ContextBinding::VarBinding {
+                            var: "x".to_owned(),
+                            ty: Ty::Int(),
+                        },
+                        ContextBinding::VarBinding {
+                            var: "xs".to_owned(),
+                            ty: Ty::Decl("ListInt".to_owned()),
+                        },
+                        ContextBinding::CovarBinding {
+                            covar: "a".to_owned(),
+                            ty: Ty::Int(),
+                        },
+                    ],
+                    rhs: Rc::new(
+                        Cut {
+                            producer: Rc::new(
+                                Variable {
+                                    var: "x".to_owned(),
+                                }
+                                .into(),
+                            ),
+                            consumer: Rc::new(
+                                Covariable {
+                                    covar: "a".to_owned(),
+                                }
+                                .into(),
+                            ),
+                        }
+                        .into(),
+                    ),
+                },
+            ],
+        }
+        .into()
+    }
+
+    #[test]
+    fn display_cocase() {
+        let result = format!("{}", example_cocase());
+        let expected =
+            "cocase { Fst(x : Int, 'a :cnt Int) => <x | 'a>, Snd() => <x | 'a> }".to_owned();
+        assert_eq!(result, expected)
+    }
+
+    #[test]
+    fn display_case() {
+        let result = format!("{}", example_case());
+        let expected =
+            "case { Nil() => <x | 'a>, Cons(x : Int, xs : ListInt, 'a :cnt Int) => <x | 'a> }"
+                .to_owned();
+        assert_eq!(result, expected)
+    }
+
+    #[test]
+    fn free_vars_cocase() {
+        let result = example_cocase().free_vars();
+        let expected = HashSet::from(["x".to_owned()]);
+        assert_eq!(result, expected)
+    }
+
+    #[test]
+    fn free_vars_case() {
+        let result = example_case().free_vars();
+        let expected = HashSet::from(["x".to_owned()]);
+        assert_eq!(result, expected)
+    }
+
+    #[test]
+    fn free_covars_cocase() {
+        let result = example_cocase().free_covars();
+        let expected = HashSet::from(["a".to_owned()]);
+        assert_eq!(result, expected)
+    }
+
+    #[test]
+    fn free_covars_case() {
+        let result = example_case().free_covars();
+        let expected = HashSet::from(["a".to_owned()]);
+        assert_eq!(result, expected)
+    }
+}
